requirements.txt
RPi.GPIO=0.7.1
python-dotenv==1.1.0
paho-mqtt==2.1.0
stmpy==0.7.5